Use single codepath for sending transaction for local and remote nodes
Previously procedure for sending transactions was different based on
which type of node was used, local (light node) or remote.

This change removes separate code path for working with light node,
and the only difference will be backend for rpc client. In case of
light node all of the communication will be done over inproc dial.

Also there is a couple of related changes in this PR:
- new EthereumTransactor that provides higher level API for working
  with ethereum network, and it is fully conformant with ethclient
- new test rpc service that improves flexibility and coverage of
  txqueue manager tests